requirements/base.txt
Django==1.6.5
bpython==0.13
django-braces==1.4.0
django-model-utils==2.0.3
logutils==0.3.3
django-annoying==0.8.0
django-mptt==0.6.1
django_mptt_admin==0.1.8
pillow==2.5.0
djangorestframework==2.3.14
South==1.0
django-simple-captcha==0.4.2
django-registration==1.0
django-suit==0.2.9
DjangoUeditor
django-taggit
django-imagekit
django-filter
